[T_vumeter] change decay to float type

+++ avidemux/qt4/ADM_userInterfaces/ADM_gui/T_vumeter.h
@@ -13,7 +13,7 @@ class  ADM_Qvumeter : public QWidget
     Q_OBJECT
 public:
     typedef struct {
-        int maxPos;
+        double maxPos;
         uint32_t maxColor;
     } peak_t;
